Add other pools
