+++ balancer-rs/src/generated_contracts/mod.rs
@@ -1,3 +1,5 @@
+pub mod liquidity_bootstrapping_pool;
+pub mod meta_stable_pool;
 pub mod simple_test_contract;
 pub mod vault;
 pub mod weighted_pool;
